Utils folder
Added the utils folder to contain functions that aren't related to one particular feature.

Updated calls to session.user._id to a wrapper to prevent breaking and added an input xss protection for future use.

Finally, updated validation to account for spaces and symbols to allow the seed file to work correctly.